actions: Split out zerver.actions.presence.
