+++ zerver/openapi/curl_param_value_generators.py
@@ -10,9 +10,10 @@ from typing import Any, Callable, Dict, List, Optional, Set, Tuple
 
 from django.utils.timezone import now as timezone_now
 
+from zerver.actions.presence import update_user_presence
 from zerver.actions.realm_linkifiers import do_add_linkifier
 from zerver.actions.realm_playgrounds import do_add_realm_playground
-from zerver.lib.actions import do_add_reaction, do_create_user, update_user_presence
+from zerver.lib.actions import do_add_reaction, do_create_user
 from zerver.lib.events import do_events_register
 from zerver.lib.initial_password import initial_password
 from zerver.lib.test_classes import ZulipTestCase
